@@ -1,6 +1,7 @@
 package coreJava.lambda;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import javax.swing.*;
 import java.time.Instant;
@@ -17,7 +18,7 @@ public class LambdaTest {
 
     @Test
     public void testLambda() {
-        var planets = new String[] { "Mercury", "Venus", "Earth", "Mars",
+        String[] planets = new String[] { "Mercury", "Venus", "Earth", "Mars",
                 "Jupiter", "Saturn", "Uranus", "Neptune" };
         System.out.println(Arrays.toString(planets));
         System.out.println("Sorted in dictionary order:");
@@ -27,7 +28,7 @@ public class LambdaTest {
         Arrays.sort(planets, Comparator.comparingInt(String::length));
         System.out.println(Arrays.toString(planets));
 
-        var timer = new Timer(1000, e -> {
+        Timer timer = new Timer(1000, e -> {
             System.out.println("Instant.ofEpochMilli(e.getWhen()) = " + Instant.ofEpochMilli(e.getWhen()));
         });
         timer.start();
